+++ bin/faucet/src/stub_rpc_api.rs
@@ -105,13 +105,6 @@ impl api_server::Api for StubRpcApi {
         unimplemented!()
     }
 
-    async fn get_account_state_delta(
-        &self,
-        _request: Request<proto::rpc_store::AccountStateDeltaRequest>,
-    ) -> Result<Response<proto::rpc_store::AccountStateDelta>, Status> {
-        unimplemented!()
-    }
-
     async fn get_account_proofs(
         &self,
         _request: Request<proto::rpc_store::AccountProofsRequest>,