+++ crates/proto/src/clients/mod.rs
@@ -33,6 +33,7 @@ use tonic::service::Interceptor;
 use tonic::service::interceptor::InterceptedService;
 use tonic::transport::{Channel, ClientTlsConfig, Endpoint};
 use tonic::{Request, Status};
+use url::Url;
 
 use crate::generated;
 
@@ -269,19 +270,16 @@ impl<State> Builder<State> {
 impl Builder<WantsTls> {
     /// Create a new strict builder from a gRPC endpoint URL such as
     /// `http://localhost:8080` or `https://api.example.com:443`.
-    pub fn new(url: impl Into<String>) -> Result<Builder<WantsTls>> {
-        let address = url.into();
-        // Basic upfront validation using `Endpoint::from_shared` path; let `Builder` do full checks
-        // later to keep logic in one place.
-        let endpoint = Endpoint::from_shared(address.clone())
-            .context("Failed to create endpoint from address")?;
+    pub fn new(url: Url) -> Builder<WantsTls> {
+        let endpoint = Endpoint::from_shared(String::from(url))
+            .expect("Url type always results in valid endpoint");
 
-        Ok(Builder {
+        Builder {
             endpoint,
             metadata_version: None,
             metadata_genesis: None,
             _state: PhantomData,
-        })
+        }
     }
 
     /// Explicitly disable TLS.