+++ crates/store/src/db/mod.rs
@@ -7,7 +7,7 @@ use miden_lib::utils::Serializable;
 use miden_node_proto::domain::account::{AccountInfo, AccountSummary, NetworkAccountPrefix};
 use miden_node_proto::generated as proto;
 use miden_objects::Word;
-use miden_objects::account::{AccountDelta, AccountId};
+use miden_objects::account::AccountId;
 use miden_objects::block::{BlockHeader, BlockNoteIndex, BlockNumber, ProvenBlock};
 use miden_objects::crypto::merkle::SparseMerklePath;
 use miden_objects::note::{NoteDetails, NoteId, NoteInclusionProof, NoteMetadata, Nullifier};
@@ -372,7 +372,8 @@ impl Db {
         .await
     }
 
-    /// Loads all the [`Note`]s matching a certain [`NoteId`] from the database.
+    /// Loads all the [`miden_objects::note::Note`]s matching a certain [`NoteId`] from the
+    /// database.
     #[instrument(level = "debug", target = COMPONENT, skip_all, ret(level = "debug"), err)]
     pub async fn select_notes_by_id(&self, note_ids: Vec<NoteId>) -> Result<Vec<NoteRecord>> {
         self.transact("note by id", move |conn| {
@@ -438,24 +439,10 @@ impl Db {
         .await
     }
 
-    /// Merges all account deltas from the DB for given account ID and block range.
-    /// Note, that `from_block` is exclusive and `to_block` is inclusive.
-    ///
-    /// Returns `Ok(None)` if no deltas were found in the DB for the specified account within
-    /// the given block range.
-    pub(crate) async fn select_account_state_delta(
-        &self,
-        account_id: AccountId,
-        from_block: BlockNumber,
-        to_block: BlockNumber,
-    ) -> Result<Option<AccountDelta>> {
-        self.transact("select account state data", move |conn| {
-            models::queries::select_account_delta(conn, account_id, from_block, to_block)
-        })
-        .await
-    }
-
-    /// Selects storage map values for syncing within a block range, with pagination support.
+    /// Selects storage map values for syncing storage maps for a specific account ID.
+    /// 
+    /// The returned values are the latest known values up to `block_to`, and no values earlier
+    /// than `block_from` are returned.
     pub(crate) async fn select_storage_map_sync_values(
         &self,
         account_id: AccountId,